requirements.txt
PySide6
PySide6-Fluent-Widgets
PyYAML>=6.0
requests>=2.31
httpx>=0.24
json-repair
Jinja2>=3.1
openpyxl>=3.1
xlrd>=2.0
pyinstaller>=6.6
nuitka>=2.3
